Moved and renamed listener_filter_fakes.h
