+++ test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.cc
@@ -1,7 +1,5 @@
 #include "test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.h"
 
-#include "common/network/utility.h"
-
 namespace Envoy {
 namespace Extensions {
 namespace ListenerFilters {
